+++ CHANGELOG.md
@@ -5,7 +5,7 @@ file is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and
 this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.22.0] - 2023-01-09
 
 ### Added
 
@@ -23,6 +23,8 @@ Versioning](https://semver.org/spec/v2.0.0.html).
     and scalability.
   - A migration function has been provided to seamlessly transition data from
     the old PostgreSQL table to RocksDB.
+- `nodes` table's fields are modified. Migration of data is supported by function
+  `migrate_0_20_to_0_22`.
 
 ### Deprecated
 
@@ -45,7 +47,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
   you create a model using `Model::from_storage`, the  `batch_info` and `scores`
   fields will now be initialized with their default values. If you previously
   relied on custom values for these fields, you will need to update your code accordingly.
-  
+
 ### Removed
 
 - `event_range` Table Removal:
@@ -565,7 +567,7 @@ leading to a more streamlined system.
 
 - An initial version.
 
-[Unreleased]: https://github.com/petabi/review-database/compare/0.21.0...main
+[0.22.0]: https://github.com/petabi/review-database/compare/0.21.0...0.22.0
 [0.21.0]: https://github.com/petabi/review-database/compare/0.20.0...0.21.0
 [0.20.0]: https://github.com/petabi/review-database/compare/0.19.0...0.20.0
 [0.19.0]: https://github.com/petabi/review-database/compare/0.18.0...0.19.0